Main Page(Landing Page),Learning path&bootcamp
main page when you want to run the program, run it Landing Page.
In this project, it is a technical project for testing the front end responsive website which can be run on the website and on mobile phones. In this project, you can move pages, swipe pages, direct divs in the same folder.